+++ aldrin/CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Remove `RequiredFieldMissing` error.
+- Rename `InvalidFunction::function()` to `id()`.
 
 ## [0.10.0] - 2024-11-26